feat: Move build_simulation_overrides into Chain
